+++ src/__tests__/MarkdownTogglableDetails.test.tsx
@@ -1,4 +1,4 @@
-import MarkdownTogglableDetails from "@/app/cards/markdown_togglableDetails";
+import MarkdownTogglableDetails from "@/app/cards/MarkdownTogglableDetails";
 import "@testing-library/jest-dom";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
